API-65: fixup rebase issues
