+++ .env.example
@@ -7,4 +7,4 @@ PGADMIN_DEFAULT_EMAIL=dev@badgehub.team
 
 NODE_ENV=development
 
-JWT_SIGNING_KEY="!!HackThePlanet!!"
+KEYCLOAK_ISSUER=https://keycloak.p1m.nl/realms/master
