+++ .env.prod.example
@@ -5,4 +5,4 @@ POSTGRES_PASSWORD=badgehub
 PGADMIN_DEFAULT_PASSWORD=badgehub
 PGADMIN_DEFAULT_EMAIL=dev@badgehub.team
 
-JWT_SIGNING_KEY="!!HackThePlanet!!"
+KEYCLOAK_ISSUER=https://keycloak.p1m.nl/realms/master
