+++ .env.test
@@ -6,4 +6,4 @@ PGADMIN_DEFAULT_PASSWORD=badgehub
 PGADMIN_DEFAULT_EMAIL=dev@badgehub.team
 DISABLE_AUTH=true
 
-JWT_SIGNING_KEY="!!HackThePlanet!!"
+KEYCLOAK_ISSUER=https://keycloak.p1m.nl/realms/master
