+++ src/config.ts
@@ -10,7 +10,6 @@ export const POSTGRES_PORT = 5432;
 export const NODE_ENV = process.env.NODE_ENV;
 export const DISABLE_AUTH = process.env.DISABLE_AUTH === "true";
 export const MAX_UPLOAD_FILE_SIZE_BYTES = 32 * 1024 * 1024; // 32 MB
-export const KEYCLOAK_ISSUER =
-  "https://lemur-11.cloud-iam.com/auth/realms/badgehub";
+export const KEYCLOAK_ISSUER = process.env.KEYCLOAK_ISSUER;
 export const KEYCLOAK_CERTS =
-  "https://lemur-11.cloud-iam.com/auth/realms/badgehub/protocol/openid-connect/certs";
+  KEYCLOAK_ISSUER + "/protocol/openid-connect/certs";
