+++ src/controllers/public-rest.ts
@@ -1,4 +1,3 @@
-import { Security, Response } from "tsoa";
 import { Controller, type TsoaResponse } from "tsoa";
 import { Get, Path, Query, Res, Route, Tags } from "tsoa";
 import { BadgeHubData } from "@domain/BadgeHubData";
@@ -26,11 +25,6 @@ import type { RevisionNumber } from "@domain/readModels/project/Version";
  * npm run swagger
  */
 
-export interface JwtError extends Error {
-  status: number;
-  message: string;
-}
-
 @Route("/api/v3")
 @Tags("public")
 export class PublicRestController extends Controller {
@@ -41,19 +35,6 @@ export class PublicRestController extends Controller {
     )
   ) {
     super();
-  }
-
-  /**
-   * Only for testing auth endpoint
-   */
-  @Response<JwtError>(403, "Forbidden") // Doesn't work
-  @Security("bearer", ["hacker"])
-  @Get("/private")
-  public async getPrivate() {
-    return Promise.resolve({
-      status: 200,
-      message: "You're visiting a private api",
-    });
   }
 
   /**
